unsubscribed all subscription on ngdestroy

+++ src/app/modules/dashboard/components/recipes/recipe-preview/recipe-preview.component.ts
@@ -3,24 +3,32 @@ import {Subject} from 'rxjs';
 import {Recipe} from '../../../models/recipe.model';
 import {Router} from '@angular/router';
 import {RecipeManager} from '../../../managers/recipe.manager';
-import {filter} from 'rxjs/operators';
+import {filter, takeUntil} from 'rxjs/operators';
 import {NotificationService} from '../../../../shared/services/notification.service';
-import {ShoppingService} from '../../../services/shopping.service';
 import {ShoppingManager} from '../../../managers/shopping.manager';
 import {FormControl, FormGroup} from '@angular/forms';
 import {Ingredient} from '../../../models/ingredient.model';
+import {UnsubscribeAbstract} from '../../../../shared/components/unsubscribe/unsubscribe.component';
+import {LoggedInUserManager} from '../../../../auth/managers/logged-in-user.manager';
+import {UserModel} from '../../../../shared/models/user.model';
 
 @Component({
   selector: 'rb-recipe-preview',
   templateUrl: './recipe-preview.component.html',
   styleUrls: ['./recipe-preview.component.scss']
 })
-export class RecipePreviewComponent implements OnInit {
+export class RecipePreviewComponent extends UnsubscribeAbstract implements OnInit {
 
   constructor(private _router: Router,
               private _notificationService: NotificationService,
               private _recipeManager: RecipeManager,
-              private _shoppingManager: ShoppingManager) { }
+              private _shoppingManager: ShoppingManager,
+              private _loggedInUserManager: LoggedInUserManager) {
+    super();
+    this._loggedInUserManager.selectLoggedInUser().pipe(takeUntil(this.destroyed$)).subscribe(loggedInUser => {
+      this.loggedInUser = loggedInUser;
+    });
+  }
 
   recipe: Recipe;
   onClose = new Subject<boolean>();
@@ -30,6 +38,7 @@ export class RecipePreviewComponent implements OnInit {
   ingredientsForm = new FormGroup(this.control);
   selectedIngredientsCount = 0;
 
+  loggedInUser: UserModel;
 
   ngOnInit() {
     this.handleActionState();
@@ -47,28 +56,44 @@ export class RecipePreviewComponent implements OnInit {
   }
 
   editRecipe() {
+    if (!this.canPerformModificationActions()) {
+      this._notificationService.show('Sorry, but only owner of this recipe can perform this operation', 'warn');
+      return;
+    }
     this.onClose.next();
     this._recipeManager.openRecipeEditForm(this.recipe);
     this._router.navigate(['dashboard/recipes/edit']);
   }
 
   deleteRecipe(event: MouseEvent) {
+    if (!this.canPerformModificationActions()) {
+      this._notificationService.show('Sorry, but only owner of this recipe can perform this operation', 'warn');
+      return;
+    }
     event.stopPropagation();
     this.deletingRecipe = true;
     this._recipeManager.deleteRecipe(this.recipe.id, this.recipe.is_private);
   }
 
   handleActionState() {
-    this._recipeManager.getActionState('recipeDeleted').pipe(filter(i => !!i)).subscribe(_ => {
+    this._recipeManager.getActionState('recipeDeleted').pipe(filter(i => !!i), takeUntil(this.destroyed$)).subscribe(_ => {
       this.deletingRecipe = false;
       this._notificationService.show('Recipe deleted successfully', 'success');
       this.onClose.next();
     });
-    this._shoppingManager.getActionState('ingredientsAdded').pipe(filter(i => !!i)).subscribe(_ => {
+    this._shoppingManager.getActionState('ingredientsAdded').pipe(filter(i => !!i), takeUntil(this.destroyed$)).subscribe(_ => {
       this.addingIngredient = false;
       this._notificationService.show('Ingredients added successfully to Shopping List', 'success');
       this.ingredientsForm.reset();
     });
+  }
+
+  canPerformModificationActions(): boolean {
+    if (this.recipe.is_private) {
+      return true;
+    } else {
+      return this.loggedInUser.id === this.recipe.userId;
+    }
   }
 
   addIngredients() {
@@ -83,5 +108,13 @@ export class RecipePreviewComponent implements OnInit {
       ingredients = [...this.recipe.ingredients];
     }
     this._shoppingManager.addIngredients(ingredients);
+  }
+
+  handleDeleteEvent(event: MouseEvent) {
+    if (!this.canPerformModificationActions()) {
+      this._notificationService.show('Sorry, but only owner of this recipe can perform this operation', 'warn');
+      event.stopPropagation();
+      return;
+    }
   }
 }
